new component TaskCardComponent/pagename/useMemo
